+++ fluidsim/solvers/sw1l/output/spatial_means.py
@@ -253,8 +253,14 @@ class SpatialMeansMSW1L(SpatialMeansBase):
 
     def load(self):
         dict_results = {"name_solver": self.output.name_solver}
+        return self._load(self.path_file, dict_results)
 
-        with open(self.path_file) as file_means:
+    @staticmethod
+    def _load(path_file, dict_results={}):
+        if not path_file.endswith(SpatialMeansMSW1L._name_file):
+            path_file = os.path.join(path_file, SpatialMeansMSW1L._name_file)
+
+        with open(path_file) as file_means:
             lines = file_means.readlines()
 
         lines_t = []